+++ libraries/bio/src/utils/macromolecule/consts.ts
@@ -31,6 +31,7 @@ export enum TAGS {
   positionNames = '.positionNames',
   positionLabels = '.positionLabels',
   regions = '.regions',
+  positionShift = '.positionShift',
 }
 
 export {TAGS as BioTags};