+++ libraries/compute-utils/function-views/src/variance-based-analysis/sa-outputs-routine.ts
@@ -29,7 +29,7 @@ export type OutputInfo = {
 
 export type SensitivityAnalysisResult = {
   funcEvalResults: DG.DataFrame,
-  funcCalls: DG.FuncCall[],
+  funcInputs: any[],
 };
 
 function getCellCoordinatesFromRow(df: DG.DataFrame, spec: OutputDataFromUI): CellCoordinates[] {