+++ libraries/compute-utils/reactive-tree-driver/index.ts
@@ -1,3 +1,4 @@
 export {mergeValidationResults} from './src/utils';
 export {IRuntimeLinkController, IRuntimeMetaController, IRuntimeValidatorController, IRuntimePipelineMutationController, INameSelectorController, IFuncallActionController} from './src/RuntimeControllers';
 export {PipelineConfiguration, Handler, Validator, MetaHandler, MutationHandler, SelectorHandler, FunccallActionHandler, PipelineExport} from './src/config/PipelineConfiguration';
+export {PipelineInstanceConfig} from './src/config/PipelineInstance';