+++ packages/DiffStudio/src/app.ts
@@ -1528,6 +1528,11 @@ export class DiffStudio {
       const script = DG.Script.create(scriptText);
       await SensitivityAnalysisView.fromEmpty(script, {
         inputsLookup: ivp.inputsLookup !== null ? ivp.inputsLookup : undefined,
+        diffGrok: {
+          ivp: ivp,
+          ivpWW: getIvp2WebWorker(ivp),
+          pipelineCreator: getPipelineCreator(ivp),
+        },
       });
     } catch (err) {
       this.processError(err);