+++ packages/Bio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Bio changelog
 
+## 2.21.5 (2025-05-12)
+
+* Shifted sequence rendering support
+
+## 2.21.4 (2025-05-12)
+
+* Support monomer renderer in viewers
+
 ## 2.21.3 (2025-05-01)
 
 * Similarity search viewer: Fix for clearing selection